pyproject.toml
[tool.poetry]
name = "agentic-rag"
version = "0.1.0"
description = ""
authors = ["Hisham Taj <MrHishamTaj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
beautifulsoup4 = "^4.14.3"
langchain = "^1.2.0"
langgraph = "^1.0.5"
langchainhub = "^0.1.21"
langchain-community = "^0.4.1"
tavily-python = "^0.7.17"
chromadb = "^1.4.0"
python-dotenv = "^1.2.1"
isort = "^7.0.0"
pytest = "^9.0.2"
langchain-chroma = "^1.1.0"
langchain-openai = "^1.1.6"
langchain-tavily = "^0.2.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
